# spatial/curve: Hilbert curves, take 2