See #1695 - this is the same code with minor modifications as per comments on the original PR.

<details><summary>Benchmarks</summary><pre>
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkHilbert$ gonum.org/v1/gonum/spatial/curve -v -count=1
goos: darwin
goarch: amd64
pkg: gonum.org/v1/gonum/spatial/curve
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
BenchmarkHilbert
BenchmarkHilbert/2D/Pos
BenchmarkHilbert/2D/Pos/Order_1
BenchmarkHilbert/2D/Pos/Order_1-8       100000000                11.84 ns/op            0 B/op           0 allocs/op
BenchmarkHilbert/2D/Pos/Order_2
BenchmarkHilbert/2D/Pos/Order_2-8       76892733         16.06 ns/op            0 B/op           0 allocs/op
BenchmarkHilbert/2D/Pos/Order_3
BenchmarkHilbert/2D/Pos/Order_3-8       55760536         20.37 ns/op            0 B/op           0 allocs/op
BenchmarkHilbert/2D/Pos/Order_4
BenchmarkHilbert/2D/Pos/Order_4-8       53285614                24.30 ns/op              0 B/op          0 allocs/op
BenchmarkHilbert/2D/Pos/Order_5
BenchmarkHilbert/2D/Pos/Order_5-8       38780760                26.66 ns/op              0 B/op          0 allocs/op
BenchmarkHilbert/2D/Pos/Order_6
BenchmarkHilbert/2D/Pos/Order_6-8       30895916                35.55 ns/op              0 B/op          0 allocs/op
BenchmarkHilbert/2D/Pos/Order_7
BenchmarkHilbert/2D/Pos/Order_7-8       26158227                42.13 ns/op              0 B/op          0 allocs/op
BenchmarkHilbert/2D/Pos/Order_8
BenchmarkHilbert/2D/Pos/Order_8-8       26913849                43.31 ns/op              0 B/op          0 allocs/op
BenchmarkHilbert/2D/Pos/Order_9
BenchmarkHilbert/2D/Pos/Order_9-8       26351476                43.45 ns/op              0 B/op          0 allocs/op
BenchmarkHilbert/2D/Pos/Order_10
BenchmarkHilbert/2D/Pos/Order_10-8              21631359                59.43 ns/op              0 B/op          0 allocs/op
BenchmarkHilbert/2D/Coord
BenchmarkHilbert/2D/Coord/Order_1
BenchmarkHilbert/2D/Coord/Order_1-8             176244006                7.655 ns/op           0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_2
BenchmarkHilbert/2D/Coord/Order_2-8             86029083                14.62 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_3
BenchmarkHilbert/2D/Coord/Order_3-8             88970055                16.94 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_4
BenchmarkHilbert/2D/Coord/Order_4-8             67142550                18.16 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_5
BenchmarkHilbert/2D/Coord/Order_5-8             54518832                20.30 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_6
BenchmarkHilbert/2D/Coord/Order_6-8             49024351                35.19 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_7
BenchmarkHilbert/2D/Coord/Order_7-8             39211636                25.72 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_8
BenchmarkHilbert/2D/Coord/Order_8-8             42119673                29.36 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_9
BenchmarkHilbert/2D/Coord/Order_9-8             34036394                39.42 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/2D/Coord/Order_10
BenchmarkHilbert/2D/Coord/Order_10-8            28563565                44.25 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos
BenchmarkHilbert/3D/Pos/Order_1
BenchmarkHilbert/3D/Pos/Order_1-8               45626558                23.72 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_2
BenchmarkHilbert/3D/Pos/Order_2-8               35097692                38.66 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_3
BenchmarkHilbert/3D/Pos/Order_3-8               24931842                53.10 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_4
BenchmarkHilbert/3D/Pos/Order_4-8               16363636                65.13 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_5
BenchmarkHilbert/3D/Pos/Order_5-8               17743531                73.46 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_6
BenchmarkHilbert/3D/Pos/Order_6-8               12450988                99.51 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_7
BenchmarkHilbert/3D/Pos/Order_7-8               10337702               134.7 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_8
BenchmarkHilbert/3D/Pos/Order_8-8                9742449               116.1 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_9
BenchmarkHilbert/3D/Pos/Order_9-8                9062575               145.2 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/3D/Pos/Order_10
BenchmarkHilbert/3D/Pos/Order_10-8               6391741               190.9 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord
BenchmarkHilbert/3D/Coord/Order_1
BenchmarkHilbert/3D/Coord/Order_1-8             79398408                20.42 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_2
BenchmarkHilbert/3D/Coord/Order_2-8             41471281                31.48 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_3
BenchmarkHilbert/3D/Coord/Order_3-8             30333662                42.67 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_4
BenchmarkHilbert/3D/Coord/Order_4-8             21393865                55.17 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_5
BenchmarkHilbert/3D/Coord/Order_5-8             17034688                73.57 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_6
BenchmarkHilbert/3D/Coord/Order_6-8             18116989                91.11 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_7
BenchmarkHilbert/3D/Coord/Order_7-8             13880876                86.80 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_8
BenchmarkHilbert/3D/Coord/Order_8-8             12011598               114.2 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_9
BenchmarkHilbert/3D/Coord/Order_9-8              9410091               111.1 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/3D/Coord/Order_10
BenchmarkHilbert/3D/Coord/Order_10-8             8406057               137.7 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos
BenchmarkHilbert/4D/Pos/Order_1
BenchmarkHilbert/4D/Pos/Order_1-8               44024707                26.23 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_2
BenchmarkHilbert/4D/Pos/Order_2-8               25374694                45.76 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_3
BenchmarkHilbert/4D/Pos/Order_3-8               23462036                70.91 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_4
BenchmarkHilbert/4D/Pos/Order_4-8               15241987                75.97 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_5
BenchmarkHilbert/4D/Pos/Order_5-8               13263390               100.6 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_6
BenchmarkHilbert/4D/Pos/Order_6-8                8060668               125.5 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_7
BenchmarkHilbert/4D/Pos/Order_7-8                8870961               130.7 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_8
BenchmarkHilbert/4D/Pos/Order_8-8                7693154               145.1 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_9
BenchmarkHilbert/4D/Pos/Order_9-8                6436678               187.7 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Pos/Order_10
BenchmarkHilbert/4D/Pos/Order_10-8               5614554               234.2 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord
BenchmarkHilbert/4D/Coord/Order_1
BenchmarkHilbert/4D/Coord/Order_1-8             70980118                16.07 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_2
BenchmarkHilbert/4D/Coord/Order_2-8             30415852                40.34 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_3
BenchmarkHilbert/4D/Coord/Order_3-8             20624990                61.09 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_4
BenchmarkHilbert/4D/Coord/Order_4-8             15923089                71.61 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_5
BenchmarkHilbert/4D/Coord/Order_5-8             13974184                88.66 ns/op            0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_6
BenchmarkHilbert/4D/Coord/Order_6-8             10822477               104.1 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_7
BenchmarkHilbert/4D/Coord/Order_7-8              9425492               157.4 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_8
BenchmarkHilbert/4D/Coord/Order_8-8              7466738               179.1 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_9
BenchmarkHilbert/4D/Coord/Order_9-8              6651352               174.4 ns/op             0 B/op              0 allocs/op
BenchmarkHilbert/4D/Coord/Order_10
BenchmarkHilbert/4D/Coord/Order_10-8             6441031               193.7 ns/op             0 B/op              0 allocs/op
PASS
ok      gonum.org/v1/gonum/spatial/curve        89.593s
</pre></details>