@firelizzard18 2024-03-03
@kortschak @vladimir-ch I'm getting back to the project that I need #1691 for, so I'm revisiting this Hilbert curve code. I addressed almost all of the remaining discussions from #1695, including the `Space` and `Curve` methods to `Coord` and `Pos`. The [remaining](https://github.com/gonum/gonum/pull/1695#discussion_r711485982) [discussions](https://github.com/gonum/gonum/pull/1695#issuecomment-888678226) are about off-nominal behavior (when the caller passes out of bounds values or equivalent).