@firelizzard18 2024-03-03
Do you think it's valuable to define a `Curve` interface in the package? For example:

```go

type Curve interface {
	Dims() []int
	Len() int
	Pos(coord []int) int
	Coord(pos int) []int
}
```