@firelizzard18 2024-03-03
I eliminated all allocations from `Pos` and all those that were avoidable from `Coord` - excepting the return value - by replacing `type ops []op` and its method `do` with a function `do2` that uses generics to avoid allocating interface values. On my laptop I got the following, before and after:

<details><summary>benchcmp ns/op</summary><pre>
benchmark                               old ns/op     new ns/op     delta
BenchmarkHilbert/2D/Pos/Order_1-8       13.5          10.3          -24.15%
BenchmarkHilbert/2D/Pos/Order_2-8       16.2          15.2          -6.06%
BenchmarkHilbert/2D/Pos/Order_3-8       19.6          18.4          -5.92%
BenchmarkHilbert/2D/Pos/Order_4-8       24.4          22.7          -6.90%
BenchmarkHilbert/2D/Coord/Order_1-8     41.3          41.1          -0.53%
BenchmarkHilbert/2D/Coord/Order_2-8     46.2          44.7          -3.22%
BenchmarkHilbert/2D/Coord/Order_3-8     47.6          49.1          +3.00%
BenchmarkHilbert/2D/Coord/Order_4-8     56.4          51.0          -9.71%
BenchmarkHilbert/3D/Pos/Order_1-8       87.9          24.0          -72.71%
BenchmarkHilbert/3D/Pos/Order_2-8       172           42.4          -75.39%
BenchmarkHilbert/3D/Pos/Order_3-8       266           54.9          -79.33%
BenchmarkHilbert/3D/Pos/Order_4-8       277           72.8          -73.71%
BenchmarkHilbert/3D/Coord/Order_1-8     82.1          90.8          +10.64%
BenchmarkHilbert/3D/Coord/Order_2-8     233           128           -45.19%
BenchmarkHilbert/3D/Coord/Order_3-8     234           118           -49.53%
BenchmarkHilbert/3D/Coord/Order_4-8     284           142           -49.93%
BenchmarkHilbert/4D/Pos/Order_1-8       110           33.1          -70.08%
BenchmarkHilbert/4D/Pos/Order_2-8       237           59.4          -74.97%
BenchmarkHilbert/4D/Pos/Order_3-8       480           75.4          -84.30%
BenchmarkHilbert/4D/Pos/Order_4-8       553           97.1          -82.43%
BenchmarkHilbert/4D/Coord/Order_1-8     265           104           -60.82%
BenchmarkHilbert/4D/Coord/Order_2-8     343           117           -65.98%
BenchmarkHilbert/4D/Coord/Order_3-8     224           134           -40.07%
BenchmarkHilbert/4D/Coord/Order_4-8     425           158           -62.75%
</pre></details>
<details><summary>benchcmp allocs</summary><pre>
benchmark                               old allocs     new allocs     delta
BenchmarkHilbert/2D/Pos/Order_1-8       0              0              +0.00%
BenchmarkHilbert/2D/Pos/Order_2-8       0              0              +0.00%
BenchmarkHilbert/2D/Pos/Order_3-8       0              0              +0.00%
BenchmarkHilbert/2D/Pos/Order_4-8       0              0              +0.00%
BenchmarkHilbert/2D/Coord/Order_1-8     1              1              +0.00%
BenchmarkHilbert/2D/Coord/Order_2-8     1              1              +0.00%
BenchmarkHilbert/2D/Coord/Order_3-8     1              1              +0.00%
BenchmarkHilbert/2D/Coord/Order_4-8     1              1              +0.00%
BenchmarkHilbert/3D/Pos/Order_1-8       2              0              -100.00%
BenchmarkHilbert/3D/Pos/Order_2-8       4              0              -100.00%
BenchmarkHilbert/3D/Pos/Order_3-8       6              0              -100.00%
BenchmarkHilbert/3D/Pos/Order_4-8       6              0              -100.00%
BenchmarkHilbert/3D/Coord/Order_1-8     2              2              +0.00%
BenchmarkHilbert/3D/Coord/Order_2-8     6              2              -66.67%
BenchmarkHilbert/3D/Coord/Order_3-8     6              2              -66.67%
BenchmarkHilbert/3D/Coord/Order_4-8     6              2              -66.67%
BenchmarkHilbert/4D/Pos/Order_1-8       2              0              -100.00%
BenchmarkHilbert/4D/Pos/Order_2-8       4              0              -100.00%
BenchmarkHilbert/4D/Pos/Order_3-8       6              0              -100.00%
BenchmarkHilbert/4D/Pos/Order_4-8       6              0              -100.00%
BenchmarkHilbert/4D/Coord/Order_1-8     4              2              -50.00%
BenchmarkHilbert/4D/Coord/Order_2-8     6              2              -66.67%
BenchmarkHilbert/4D/Coord/Order_3-8     6              2              -66.67%
BenchmarkHilbert/4D/Coord/Order_4-8     8              2              -75.00%
</pre></details>
<details><summary>benchcmp bytes</summary><pre>
benchmark                               old bytes     new bytes     delta
BenchmarkHilbert/2D/Pos/Order_1-8       0             0             +0.00%
BenchmarkHilbert/2D/Pos/Order_2-8       0             0             +0.00%
BenchmarkHilbert/2D/Pos/Order_3-8       0             0             +0.00%
BenchmarkHilbert/2D/Pos/Order_4-8       0             0             +0.00%
BenchmarkHilbert/2D/Coord/Order_1-8     16            16            +0.00%
BenchmarkHilbert/2D/Coord/Order_2-8     16            16            +0.00%
BenchmarkHilbert/2D/Coord/Order_3-8     16            16            +0.00%
BenchmarkHilbert/2D/Coord/Order_4-8     16            16            +0.00%
BenchmarkHilbert/3D/Pos/Order_1-8       32            0             -100.00%
BenchmarkHilbert/3D/Pos/Order_2-8       64            0             -100.00%
BenchmarkHilbert/3D/Pos/Order_3-8       96            0             -100.00%
BenchmarkHilbert/3D/Pos/Order_4-8       96            0             -100.00%
BenchmarkHilbert/3D/Coord/Order_1-8     48            48            +0.00%
BenchmarkHilbert/3D/Coord/Order_2-8     112           48            -57.14%
BenchmarkHilbert/3D/Coord/Order_3-8     112           48            -57.14%
BenchmarkHilbert/3D/Coord/Order_4-8     112           48            -57.14%
BenchmarkHilbert/4D/Pos/Order_1-8       32            0             -100.00%
BenchmarkHilbert/4D/Pos/Order_2-8       64            0             -100.00%
BenchmarkHilbert/4D/Pos/Order_3-8       96            0             -100.00%
BenchmarkHilbert/4D/Pos/Order_4-8       96            0             -100.00%
BenchmarkHilbert/4D/Coord/Order_1-8     96            64            -33.33%
BenchmarkHilbert/4D/Coord/Order_2-8     128           64            -50.00%
BenchmarkHilbert/4D/Coord/Order_3-8     128           64            -50.00%
BenchmarkHilbert/4D/Coord/Order_4-8     160           64            -60.00%
</pre></details>