@firelizzard18 2024-03-03
I replaced `CoordND(pos int) [N]int` with `WriteCoordTo(pos int, v []int)` and updated the benchmark to use that. That eliminated *all* allocations in the benchmark and improved the performance of `WriteCoordTo` (vs `CoordND`) by another ~50%. I'm doing this because my application will be making heavy use of the R-tree (once that's implemented) and the performance of lookups will meaningfully impact the user experience.