@firelizzard18 2024-03-03
WDYT about making these generic? E.g. `type HilbertND[V constraints.Integer] struct{ Order V }`. That would allow the caller to use any integer type, including uint64 if 32-bit overflow is a concern.